feat: use ExecutionStatus from SDK (#10079)
Please write a summary of your changes and why you made them.

Fixes https://github.com/iotaledger/iota/issues/9843

- [ ] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [ ] Patch-specific tests (correctness, functionality coverage)
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have checked that new and existing unit tests pass locally with
my changes

+++ crates/iota-indexer/src/apis/write_api.rs
@@ -294,7 +294,7 @@ impl WriteApi {
 
                 let mut error = ExecutionError::new(exec_err.into(), source);
                 if let Some(command_index) = execution_error.command_index {
-                    error = error.with_command_index(command_index as usize);
+                    error = error.with_command_index(command_index);
                 }
                 Ok(error.to_string())
             })
